seeed_format add the 1st zip file (not ready to submit)
Add result of script run, the symlink files and .zip file

Pending:

Seeed Studio requires silkscreen overlay (.G[BT]O) files,
even if there is no silkscreen required.

In which case, since the silkscreen is a negative,
a filled board outline can be submitted.
However it'd be nicer to have an overlay, right?